updated ArrayList<String> to String[]
